removing PII and moving fixtures
