+++ src/routes/v1/apply/accreditation.test.js
@@ -2,7 +2,7 @@ import {
   LOGGING_EVENT_ACTIONS,
   LOGGING_EVENT_CATEGORIES
 } from '../../../common/enums/event.js'
-import accreditationData from '../../../../fixtures/accreditation.json'
+import accreditationData from '../../../test/fixtures/accreditation.json'
 
 const mockLoggerInfo = vi.fn()
 const mockLoggerError = vi.fn()
