+++ src/test/fixtures/registration.json
